Feat: Cria novo service e controller para métodos do carrinho
